+++ examples/queue_kkt_example_document.php
@@ -303,7 +303,7 @@ foreach ((array)$documentArray['items'] as $product) {
     $productDto = new Product(
         (string)$product['name'],
         $money,
-        (int)$product['quantity'],
+        (float)$product['quantity'],
         $priceSum,
         (string)$product['tax'],
         $taxSum,
